api/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
Pillow==10.4.0
torch==2.4.0
torchvision==0.19.0
numpy